handle missing required properties
